refactor(accounts): rename company CRUD and dialog modules
P1-6 slice 3: company-crud-impl, contact/stakeholder dialogs → account-*; symbols AccountContactDialog / AccountStakeholderDialog.
