+++ docs/tech-debt-inventar.md
@@ -77,7 +77,7 @@ Verweis: [arbeitspaket-logging-error-e6.md](./arbeitspaket-logging-error-e6.md),
 | P1-3  | God-Files                                                   | große Actions/UI-Dateien                                                        | Sliceweise splitten                                                            | L       | E5 — **✅ Welle 2026-08-04** (Overview ~719, Share-Link ~420, Feed ~429, Match ~270)     |
 | P1-4  | `companyFromJoin` 3× + inline in Deals                      | Duplikat                                                                        | Eine Shared-Helper-Funktion                                                    | S       | neu — ✅ `lib/accounts/company-from-join.ts`                                             |
 | P1-5  | `normalizeDealStatus` 3× (actions, request, market-signals) | Drift-Risiko                                                                    | Eine Funktion in `lib/deals/`                                                  | S       | neu — ✅ `lib/deals/normalize-deal-status.ts`                                            |
-| P1-6  | Accounts Naming                                             | Route `accounts/`, Code `company*` (~18 Dateien)                                | Schrittweise Rename (DB `companies` ok)                                        | L       | Welle 5 — Slice 1 ✅ detail; Slice 2: `companies-grid*` → `accounts-grid*` |
+| P1-6  | Accounts Naming                                             | Route `accounts/`, Code `company*` (~18 Dateien)                                | Schrittweise Rename (DB `companies` ok)                                        | L       | Welle 5 — Slice 1–3 ✅ (detail/grid/crud); Rest: Lib `company-*` |
 | P1-7  | PDF/Extract 5 Module, 2 Einstiege                           | `document-extraction` vs `extract-rfp-plain-text` → …                           | Facade(s) dokumentieren/konsolidieren                                          | M       | neu — ✅ siehe Abschnitt unten                                                           |
 | P1-8  | Match Lib-Split                                             | `lib/match/*` + top-level `lib/match-*.ts` + Orchestrator in `library/match.ts` | Klare Schicht; Typen nicht aus Dashboard importieren                           | M       | E5 — ✅ top-level → `lib/match/`; Orchestrator ~270                                      |
 | P1-9  | Dual Form-Schicht References                                | Re-export New → `lib/references/reference-form/*`; Fields noch unter dashboard  | Lib von Dashboard entkoppeln                                                   | M       | E5 — ✅ Typen + Fields in lib; CreateContactDialog bleibt in app                         |
